backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
transformers==4.45.2
sentence-transformers==3.1.1
chromadb==0.5.5
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
tqdm==4.66.5
sacrebleu==2.4.3
rouge-score==0.1.2
nltk==3.9.1
google-generativeai==0.8.3
detoxify==0.5.2
datasets==2.19.1
# Generation / evaluation
evaluate==0.4.3
bert-score==0.3.13
# Torch versions vary by CUDA; install suitable version if this fails
# Example CPU-only fallback:
# torch==2.4.1 --extra-index-url https://download.pytorch.org/whl/cpu
python-dotenv==1.0.1
accelerate>=0.26.0
